Small SRP formatting updates

+++ exercises/single_responsibility/srp_orders/CustomerOrderData.java
@@ -1,15 +1,30 @@
+import java.math.BigDecimal;
+
 public class CustomerOrderData {
+
     private String orderId;
     private String itemName;
-    private double amount;
+    private BigDecimal amount;
 
-    public CustomerOrderData(String orderId, String itemName, double amount) {
+    public CustomerOrderData(
+        String orderId,
+        String itemName,
+        BigDecimal amount
+    ) {
         this.orderId = orderId;
         this.itemName = itemName;
         this.amount = amount;
     }
 
-    public String getOrderId() { return orderId; }
-    public String getItemName() { return itemName; }
-    public double getAmount() { return amount; }
+    public String getOrderId() {
+        return orderId;
+    }
+
+    public String getItemName() {
+        return itemName;
+    }
+
+    public BigDecimal getAmount() {
+        return amount;
+    }
 }
